+++ apps/prairielearn/src/pages/instructorQuestionSettings/instructorQuestionSettings.html.tsx
@@ -13,7 +13,6 @@ import { TagDescription } from '../../components/TagDescription.js';
 import { TopicBadgeHtml } from '../../components/TopicBadge.js';
 import { TopicDescription } from '../../components/TopicDescription.js';
 import { compiledScriptTag, nodeModulesAssetPath } from '../../lib/assets.js';
-import { config } from '../../lib/config.js';
 import {
   AssessmentSchema,
   AssessmentSetSchema,
@@ -263,7 +262,12 @@ export function InstructorQuestionSettings({
                   ${shouldShowAssessmentsList
                     ? html`<tr>
                         <th class="align-middle">Assessments</th>
-                        <td>${AssessmentBadges({ assessmentsWithQuestion, resLocals })}</td>
+                        <td>
+                          ${AssessmentBadges({
+                            assessmentsWithQuestion,
+                            courseInstanceId: resLocals.course_instance.id,
+                          })}
+                        </td>
                       </tr>`
                     : ''}
                 </tbody>
@@ -900,7 +904,6 @@ function DeleteQuestionModal({
                     <div class="h6">${a_with_q.short_name} (${a_with_q.long_name})</div>
                     ${a_with_q.assessments.map((assessment) =>
                       AssessmentBadgeHtml({
-                        plainUrlPrefix: config.urlPrefix,
                         courseInstanceId: a_with_q.course_instance_id,
                         assessment,
                       }),
@@ -1006,13 +1009,11 @@ function QuestionSharing({
 
 function AssessmentBadges({
   assessmentsWithQuestion,
-  resLocals,
+  courseInstanceId,
 }: {
   assessmentsWithQuestion: SelectedAssessments[];
-  resLocals: Record<string, any>;
+  courseInstanceId: string;
 }) {
-  const courseInstanceId = resLocals.course_instance.id;
-
   const assessmentsInCourseInstance = assessmentsWithQuestion.find((a) =>
     idsEqual(a.course_instance_id, courseInstanceId),
   );